// src/lib/weebdex/model/relation.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * WeebDex API
 * WeebDex is an ad-free, scanlator-first online manga reader providing high-quality, zero-compression images. <br /><br />
Contact: [delta@weebdex.org](mailto:delta@weebdex.org) or join our [Discord Server](https://discord.gg/syp3a6Eucq).<br /><br />
All API endpoints have a global rate limit of 5 requests per second per IP. <br />
To avoid future issues, include the Origin: https://weebdex.org and Referer: https://weebdex.org/ headers when making API requests.<br/>
 * OpenAPI spec version: 1.2.0
 */
import type { Demographic } from "./demographic";
import type { RelationAltTitles } from "./relationAltTitles";
import type { RelationRelationships } from "./relationRelationships";
import type { RelationType } from "./relationType";

export interface Relation {
  alt_titles?: RelationAltTitles;
  content_rating?: string;
  created_at?: string;
  deleted_at?: string;
  demographic?: Demographic;
  description?: string;
  id?: string;
  language?: string;
  last_chapter?: string;
  last_volume?: string;
  locked?: boolean;
  published_at?: string;
  relationships?: RelationRelationships;
  state?: string;
  status?: string;
  title?: string;
  type?: RelationType;
  updated_at?: string;
  version?: number;
  year?: number;
}