// src/lib/weebdex/model/chapterAggregateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Suicaodex API
 * WeebDex API đè tem, ăn quả úp bô nên không backup kịp data.

Phải gì ạ? Phải chịuuuuuu....
 * OpenAPI spec version: 1.0.0
 */
import type { ChapterAggregate } from "./chapterAggregate";
import type { ChapterAggregateResponseGroupsItem } from "./chapterAggregateResponseGroupsItem";

export interface ChapterAggregateResponse {
  chapters: ChapterAggregate[];
  groups: ChapterAggregateResponseGroupsItem[];
  languages: string[];
}